requirements.txt
opencv-python==4.11.0.86
numpy==1.23.5
tensorflow==2.10.0
pillow==11.1.0
scikit-learn==1.6.1